Refactors Context to be in its own file
