+++ phaser/phase.py
@@ -4,7 +4,8 @@ import pandas as pd
 import logging
 
 from .column import make_strict_name, Column
-from .pipeline import Context, DropRowException, DataException, PhaserError
+from .pipeline import DropRowException, DataException, PhaserError
+from . import Context
 from .records import Records, Record
 from .steps import ROW_STEP, BATCH_STEP, CONTEXT_STEP, PROBE_VALUE, row_step, DATAFRAME_STEP
 from .constants import *
